auth_service/requirements.txt
aiokafka
fastapi
passlib
python-jose
jwt
uvicorn
pydantic[email]
pydantic-settings
requests